customer-service/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
marshmallow==3.20.1

# Optional for testing
pytest==7.4.0
pytest-flask==1.3.0

# Optional for migrations
Flask-Migrate==4.0.5

# Optional for .env config loading
python-dotenv==1.0.1

cx_Oracle==8.3.0